Improve in the get_versions method
Added indentation to the returned JSON's

Rewrite the get_version to simplify it and make it more readable

Added extractor module that implement the extraction function from a html element based on the what you need, the names of functions make it clear what each function does
